+++ com.unity.render-pipelines.high-definition/Editor/AssetProcessors/MaterialPostProcessor.cs
@@ -365,38 +365,23 @@ namespace UnityEditor.Rendering.HighDefinition
 
                 void AddDiffusionProfileToSettings(string propName)
                 {
-                    if (Application.isBatchMode || HDRenderPipelineGlobalSettings.instance == null
-                        || HDRenderPipelineGlobalSettings.instance.diffusionProfileSettingsList == null) return;
-
-                    bool diffusionProfileCanBeAdded = HDRenderPipelineGlobalSettings.instance.diffusionProfileSettingsList.Length < 15;
-                    DiffusionProfileSettings diffusionProfile = null;
+                    if (Application.isBatchMode) return;
 
                     if (material.HasProperty(propName))
                     {
                         var diffusionProfileAsset = material.GetVector(propName);
                         string guid = HDUtils.ConvertVector4ToGUID(diffusionProfileAsset);
-                        diffusionProfile = AssetDatabase.LoadAssetAtPath<DiffusionProfileSettings>(AssetDatabase.GUIDToAssetPath(guid));
+                        var assetPath = AssetDatabase.GUIDToAssetPath(guid);
 
+                        var diffusionProfile = AssetDatabase.LoadAssetAtPath<DiffusionProfileSettings>(assetPath);
                         if (diffusionProfile != null && !HDRenderPipelineGlobalSettings.instance.diffusionProfileSettingsList.Any(d => d == diffusionProfile))
-                        {
-                            string materialName = material.name;
-                            string diffusionProfileName = diffusionProfile.name;
-
-                            if (!diffusionProfileCanBeAdded)
-                                Debug.LogWarning("There is no space in the global settings to add the diffusion profile " + diffusionProfileName);
-                            else if ((!Application.isBatchMode) &&
-                                     (EditorUtility.DisplayDialog("Diffusion Profile Import",
-                                         "A Material (" + materialName + ") is being imported with a diffusion profile (" + diffusionProfileName + ") not already added to the HDRP Global Settings.\n If the Diffusion Profile is not referenced in the global settings, HDRP cannot use it.\nDo you want to add the diffusion profile to the HDRP Global Settings asset?", "Yes", "No")))
-                            {
-                                diffusionProfileCanBeAdded = HDRenderPipelineGlobalSettings.instance.AddDiffusionProfile(diffusionProfile);
-                            }
-                        }
+                            MaterialDiffusionProfileReferences.RequireDiffusionProfile(assetPath);
                     }
                 }
 
                 AddDiffusionProfileToSettings("_DiffusionProfileAsset");
 
-                // Special Eye case that uses a node with diffusion profiles.
+                // ShaderGraph can reference custom diffusion profiles.
                 if (material.shader.IsShaderGraphAsset())
                 {
                     var matProperties = MaterialEditor.GetMaterialProperties(new UnityEngine.Object[] { material });