+++ com.unity.render-pipelines.high-definition/CHANGELOG.md
@@ -78,6 +78,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Added a minimum motion vector length to the motion vector debug view.
 - Added a better support for LODs in the ray tracing acceleration structure.
 - Added a property on the HDRP asset to allow users to avoid ray tracing effects running at too low percentages (case 1342588).
+- Added a window to choose missing diffusion profiles to add to global settings when importing materials.
 
 ### Fixed
 - Fixed Intensity Multiplier not affecting realtime global illumination.